[ feature ] contents image hooks
